refactoring and reducing complexity
